+++ app/services/home/index_data.rb
@@ -1,6 +1,6 @@
 module Home
   class IndexData
-    Result = Struct.new(:featured_projects, :featured_blog_posts, :featured_videos, keyword_init: true)
+    Result = Struct.new(:featured_projects, :featured_videos, keyword_init: true)
 
     def self.call
       new.call
@@ -9,7 +9,6 @@ module Home
     def call
       Result.new(
         featured_projects: Project.where(featured: true),
-        featured_blog_posts: BlogPost.where(featured: true),
         featured_videos: Video.where(featured: true)
       )
     end